+++ modules/alloy/src/main/ts/ephox/alloy/ui/types/MenuTypes.ts
@@ -1,3 +1,6 @@
+import { Optional } from '@ephox/katamari';
+import { SugarElement } from '@ephox/sugar';
+
 import { AlloyBehaviourRecord } from '../../api/behaviour/Behaviour';
 import { AlloyComponent } from '../../api/component/ComponentApi';
 import { SketchBehaviours } from '../../api/component/SketchBehaviours';
@@ -19,6 +22,7 @@ export interface MenuGridMovementSpec {
 export interface MenuMatrixMovementSpec {
   mode: 'matrix';
   rowSelector: string;
+  previousSelector?: (comp: AlloyComponent) => Optional<SugarElement<HTMLElement>>;
 }
 
 export interface MenuNormalMovementSpec {
@@ -42,6 +46,7 @@ export interface MenuMatrixMovement {
   mode: 'matrix';
   config: (detail: MenuDetail, movementInfo: MenuMovement) => MatrixConfigSpec;
   rowSelector: string;
+  previousSelector: (comp: AlloyComponent) => Optional<SugarElement<HTMLElement>>;
 }
 
 export interface MenuNormalMovement {